[ring3.visuals] update

+++ Ring2/Ring2.Library/Ring2/Library/ByteBufferReader.c
@@ -61,7 +61,7 @@ size_t Machine_ByteBufferReader_getPosition(Machine_ByteBufferReader* self) {
 }
 
 void Machine_ByteBufferReader_read(Machine_ByteBufferReader* self, char* p, size_t n, size_t* m) {
-  size_t l = Ring2_ByteBuffer_getNumberOfBytes(self->byteBuffer);
+  int64_t l = Ring2_ByteBuffer_getNumberOfBytes(self->byteBuffer);
   l = l < n ? l : n;
   Ring1_Memory_copyFast(p, Ring2_ByteBuffer_getBytes(self->byteBuffer) + self->position, l);
   *m = l;
